PullRequest: graalpython/3717

+++ mx.graalpython/mx_graalpython.py
@@ -382,6 +382,17 @@ def punittest(ars, report=False):
     Pass --regex to further filter the junit and TSK tests. GraalPy tests are always run in two configurations:
     with language home on filesystem and with language home served from the Truffle resources.
     """
+    path = os.environ.get("PATH", "")
+    if mx.is_linux() and not shutil.which("patchelf"):
+        venv = Path(SUITE.get_output_root()).absolute() / "patchelf-venv"
+        path += os.pathsep + str(venv / "bin")
+        if not shutil.which("patchelf", path=path):
+            mx.log(f"{time.strftime('[%H:%M:%S] ')} Building patchelf-venv with {sys.executable}... [patchelf not found on PATH]")
+            t0 = time.time()
+            subprocess.check_call([sys.executable, "-m", "venv", str(venv)])
+            subprocess.check_call([str(venv / "bin" / "pip"), "install", "patchelf"])
+            mx.log(f"{time.strftime('[%H:%M:%S] ')} Building patchelf-venv with {sys.executable}... [duration: {time.time() - t0}]")
+
     args = [] if ars is None else ars
     @dataclass
     class TestConfig:
@@ -420,11 +431,14 @@ def punittest(ars, report=False):
     if '--regex' not in args:
         async_regex = ['--regex', r'com\.oracle\.graal\.python\.test\.integration\.advanced\.AsyncActionThreadingTest']
         configs.append(TestConfig("async", vm_args + ['-Dpython.AutomaticAsyncActions=false', 'com.oracle.graal.python.test', 'org.graalvm.python.embedding.test'] + async_regex + args, True, False))
+    else:
+        skip_leak_tests = True
 
     for c in configs:
         mx.log(f"Python JUnit tests configuration: {c}")
         PythonMxUnittestConfig.useResources = c.useResources
-        mx_unittest.unittest(c.args, test_report_tags=({"task": f"punittest-{c.identifier}-{'w' if c.useResources else 'wo'}-resources"} if c.reportConfig else None))
+        with set_env(PATH=path):
+            mx_unittest.unittest(c.args, test_report_tags=({"task": f"punittest-{c.identifier}-{'w' if c.useResources else 'wo'}-resources"} if c.reportConfig else None))
 
     if skip_leak_tests:
         return
@@ -2682,14 +2696,15 @@ class GraalpythonProject(mx.ArchivableProject):
 orig_clean = mx.command_function("clean")
 def python_clean(args):
     orig_clean(args)
-    count = 0
-    for path in os.walk(SUITE.dir):
-        for file in glob.iglob(os.path.join(path[0], '*.pyc')):
-            count += 1
-            os.remove(file)
+    if not args:
+        count = 0
+        for path in os.walk(SUITE.dir):
+            for file in glob.iglob(os.path.join(path[0], '*.pyc')):
+                count += 1
+                os.remove(file)
 
-    if count > 0:
-        print('Cleaning', count, "`*.pyc` files...")
+        if count > 0:
+            print('Cleaning', count, "`*.pyc` files...")
 
 def update_hpy_import_cmd(args):
     """Update our import of HPy sources."""
